Update Ruby version to 3.2.8 in  .ruby-version (#240)
This PR updates the Ruby version from 3.1.3 to 3.2.8 in `.ruby-version`.

## Changes
- Updated `.ruby-version` to specify Ruby 3.2.8

## Reason for Update
Ruby 3.1 has reached End of Life (EOL) and is no longer receiving security updates or maintenance. To ensure our application remains secure and supported, we need to upgrade to a currently maintained Ruby version.

Ruby 3.2.8 is a stable release in the 3.2 series, which is currently in the security maintenance phase and will be supported until March 2025.

## Reference
- Ruby maintenance branches: https://www.ruby-lang.org/en/downloads/branches/